separate keyvalue and model structures
